pyproject.toml
[project]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = "Contacts REST API with Auth, Cloudinary and Email verification"
authors = [
    {name = "Vladyslav Popov", email = "vlad_np@ukr.net"}
]
# Важливо: додаємо <4.0, щоб задовольнити вимоги бібліотеки slowapi
requires-python = ">=3.13,<4.0"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
fastapi = ">=0.115.0"
uvicorn = ">=0.30.0"
sqlalchemy = ">=2.0.0"
psycopg2-binary = ">=2.9.9"
pydantic = {extras = ["email"], version = ">=2.0.0"}
pydantic-settings = ">=2.0.0"
python-jose = {extras = ["cryptography"], version = ">=3.3.0"}
passlib = {extras = ["bcrypt"], version = "==1.7.4"}
bcrypt = "==3.2.2"
python-multipart = ">=0.0.9"
slowapi = ">=0.1.9"
fastapi-mail = ">=1.4.1"
cloudinary = ">=1.40.0"
redis = "^7.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "sphinx (>=9.1.0,<10.0.0)",
    "sphinx-rtd-theme (>=3.1.0,<4.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
